Introduce a common base class `APIError` for all our exceptions.
